+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 - Update `wide` to `0.6.x`
 - Add `Rotor3::into_angle_plane()`
+- Add `Rotor3::into_quaternion_array` and `Rotor3::from_quaternion_array`
 
 ## 0.7.5
 - Add `Mat4::extract_translation`, `Mat4::extract_rotation` and `Mat4::into_isometry`.